#86et9akmr - feat: restructure Profile and Subscription navigation to Settings
