+++ frontend/components/partial/ProfileButton.vue
@@ -60,6 +60,6 @@
     }
 
     function goToMembership() {
-        router.push('/profile/subscription');
+        router.push('/settings/subscription');
     }
 </script>
